Split project.py into a package
